.claude/skills/visual-style-extractor/scripts/requirements.txt
scenedetect[opencv]>=0.6
imagededup>=0.3
Pillow>=10.0
numpy>=1.24
webvtt-py>=0.5
pysrt>=1.1
